Add code to handle auto-incrementing ID fields.
Some tricky transaction/concurrency stuff required
to implement a unique auto-incrementing ID field
in Google Cloud Datastore, because the database
does not support unique constraints.  We solve the
problem by performing an atomic "Get or Add", and
throw an error if the Get succeeds.  We should handle
this error and retry--not yet sure whether to do that
in the persistence layer (probably) or application logic.

Signed-off-by: Daniel Harms <jdharms@gmail.com>