Allows Polls to track missing voters
Some broad refactoring to make the Poll entity include a
"missing voters" field that will allow users to quickly query for
which voters missed a given poll.  This field is updated during
the following operations to keep consistency:

* Create Poll
* Create Ballot
* Delete Ballot
* Edit User

Notably, if a Poll is edited (for example, to change the close time),
the missing voters list isn't updated.

I plan on adding endpoints to allow admins to manually adjust the
missing voters for a poll in the near future.

Signed-off-by: Daniel Harms <jdharms@gmail.com>